Project

+++ Src/JavaSpringBoot/EcommerceBookSales/src/main/java/com/example/ecommercebooksales/entity/CartItem.java
@@ -13,7 +13,7 @@ import lombok.Setter;
 public class CartItem {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer cartItemId;
+    private Long cartItemId;
 
     @ManyToOne
     @JoinColumn(name = "cart_id")
